Refactor global functions to PascalCase naming convention and replace iostream with SDL_Log
- Rename utility functions to PascalCase: `LoadWaveRelative`, `LoadSpriteDefinitions`, `ResolveRelativeToExe`, `ProbabilityList`, and `LoadTexture`
- Replace `std::cerr`/`std::cout` with SDL logging functions (`SDL_Log`, `SDL_LogError`, `SDL_LogTrace`)
- Change timing types from `double` to `float` for `AnimatedSprite` and `SpriteRenderer`
- Add `LoadTexture` helper for BMP/PNG loading with proper error logging
- Add `getPlayTime()` method to `AudioDataBuffer` for calculating audio duration
- Update `AudioStream` to support optional clear parameter in `put`/`emplace` methods
- Add audio device initialization and stream binding logging
- Add enumeration of available render drivers on startup
- Fix `FontRenderer` text scaling and measurement calculations
- Remove unused `m_hasData` tracking from `AudioStream`
- Update include order to group system, library, and local headers

+++ sdlc/inc/AnimatedSprite.hpp
@@ -4,7 +4,6 @@
 #include <SDL3/SDL_rect.h>
 
 #include <vector>
-#include <cstdint>
 #include <functional>
 
 namespace sdlc
@@ -35,11 +34,11 @@ public:
     AnimatedSprite(const AnimatedSprite &other);
     AnimatedSprite(AnimatedSprite &&other) noexcept;
 
-    void render(SDL_Renderer* renderer, double deltaTime) override;
+    void render(SDL_Renderer* renderer, float deltaTime) override;
     int addFrame(const SDL_FRect& rect);
     int addFrames(const SpriteImageDistribution& dist);
-    void setDuration(double time);
-    void setFPS(double fps);
+    void setDuration(float time);
+    void setFPS(float fps);
     void setFrame(unsigned int number);
     void play(int frameCount = -1);
     void pause();
@@ -61,8 +60,8 @@ private:
     int m_frameCount{0};
     bool m_running{false};
     bool m_repeat{true};
-    double m_time{0};
-    double m_interval{1};
+    float m_time{0};
+    float m_interval{1};
     int m_currentFrame{0};
     Callback m_callback;
     std::vector<SDL_FRect> m_frames;
